+++ Cargo.toml
@@ -26,7 +26,7 @@ exclude = [
 features = [ "std", "backtraces", "futures", "guide" ]
 
 [features]
-default = ["std"]
+default = ["std", "rust_1_61"]
 
 # Implement the `std::error::Error` trait.
 std = []
