--- compatibility-tests/v1_61/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "v1_61"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-snafu = { path = "../..", features = ["rust_1_61"] }
